+++ MochiKit/DOM.js
@@ -431,6 +431,11 @@ MochiKit.DOM.createDOM = function (name, attrs/*, nodes... */) {
     var elem;
     var self = MochiKit.DOM;
     if (typeof(name) == 'string') {
+        // Internet Explorer is dumb
+        if (attrs && "name" in attrs && !self.attributeArray.compliant) {
+            // http://msdn.microsoft.com/workshop/author/dhtml/reference/properties/name_2.asp
+            name = '<' + name + ' name="' + self.escapeHTML(name) + '">';
+        }
         elem = self._document.createElement(name);
     } else {
         elem = name;